pyproject.toml
[project]
name = "resume-critiquer-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sniffio>=1.3.1",
    "streamlit>=1.49.1",
]
